+++ web/documentserver-example/go/config/models.go
@@ -21,10 +21,12 @@ type Extensions struct {
 	Viewed    []string `json:"viewed"`
 	Edited    []string `json:"edited"`
 	Converted []string `json:"converted"`
+	Filled    []string `json:"filled"`
 }
 
 type ExtensionTypes struct {
 	Spreadsheet  []string `json:"spreadsheet"`
 	Presentation []string `json:"presentation"`
 	Document     []string `json:"document"`
+	Pdf          []string `json:"pdf"`
 }