+++ web/documentserver-example/go/server/server.go
@@ -90,6 +90,7 @@ func (srv *Server) configureRouter() *mux.Router {
 	r.HandleFunc("/", srv.ServerAPI.Index).Methods(http.MethodGet)
 	r.HandleFunc("/editor", srv.ServerAPI.Editor).Methods(http.MethodGet)
 	r.HandleFunc("/callback", srv.ServerAPI.Callback).Methods(http.MethodPost)
+	r.HandleFunc("/remove", srv.ServerAPI.Remove).Methods(http.MethodGet)
 	r.HandleFunc("/remove", srv.ServerAPI.Remove).Methods(http.MethodDelete)
 	r.HandleFunc("/upload", srv.ServerAPI.Upload).Methods(http.MethodPost)
 	r.HandleFunc("/convert", srv.ServerAPI.Convert).Methods(http.MethodPost)
@@ -101,6 +102,8 @@ func (srv *Server) configureRouter() *mux.Router {
 	r.HandleFunc("/rename", srv.ServerAPI.Rename).Methods(http.MethodPost)
 	r.HandleFunc("/historyObj", srv.ServerAPI.HistoryObj).Methods(http.MethodPost)
 	r.HandleFunc("/restore", srv.ServerAPI.Restore).Methods(http.MethodPut)
+	r.HandleFunc("/formats", srv.ServerAPI.Formats).Methods(http.MethodGet)
+	r.HandleFunc("/forgotten", srv.ServerAPI.Forgotten).Methods(http.MethodGet, http.MethodDelete)
 
 	return r
 }